// src/models/medical_alerts/entity/medical_alerts.enum.ts
export enum ResolutionType {
  HOURLY = 'hourly',
  DAILY = 'daily',
}

export enum RiskStratification {
  BINARY = 'binary',
  TRAFFIC = 'traffic',
  CUSTOM = 'custom',
}

export enum RiskReadingType {
  CONSECUTIVE = 'consecutive',
  OUT_OF = 'out_of',
}

export enum RiskLevel {
  GREEN = 'green',
  AMBER = 'amber',
  RED = 'red',
}